fix(ci): line-limit guardrail - GC source and tests exceeded 1,500 lines
