+++ tests/unit/service/test_gc_parts/test_gc_part_005.py
@@ -1,0 +1,205 @@
+from __future__ import annotations
+
+from collections.abc import AsyncIterator
+from datetime import UTC, datetime, timedelta
+from pathlib import Path
+from unittest.mock import AsyncMock, patch
+
+import pytest
+from sqlalchemy import update
+from sqlalchemy.ext.asyncio import AsyncEngine, AsyncSession, async_sessionmaker
+
+from awf.db.enums import WorkspaceStatus
+from awf.db.models import Workspace
+from awf.db.repositories import (
+    SecretLeaseIssue,
+    SecretLeaseRepository,
+    WorkspaceRepository,
+)
+from awf.db.session import make_session_factory
+from awf.runtime.inspection import RuntimeSnapshot
+from awf.service.gc import (
+    WorkspaceGCWorktreeRemoveResult,
+    plan_terminal_workspace_gc,
+)
+
+"""Terminal workspace filesystem GC tests."""
+
+
+@pytest.fixture(autouse=True)
+def _mock_default_worktree_remover():
+    with patch(
+        "awf.service.gc._default_worktree_remover",
+        new=AsyncMock(
+            return_value=WorkspaceGCWorktreeRemoveResult(
+                status="succeeded",
+                reason_code="WORKTREE_REMOVE_SUCCEEDED",
+            )
+        ),
+    ):
+        yield
+
+
+@pytest.fixture
+async def session_factory(
+    engine: AsyncEngine,
+) -> AsyncIterator[async_sessionmaker[AsyncSession]]:
+    yield make_session_factory(engine)
+
+
+async def _workspace(
+    session_factory: async_sessionmaker[AsyncSession],
+    *,
+    status: WorkspaceStatus,
+    updated_at: datetime,
+    title: str = "gc candidate",
+    compose_file_path: str | None = None,
+    pr: bool = False,
+    pr_merge_sha: str | None = None,
+) -> str:
+    async with session_factory() as session:
+        workspace = await WorkspaceRepository(session).create(
+            repo_url="git@github.com:example/repo.git",
+            branch_base="development",
+            task_title=title,
+            task_prompt="p",
+            agent="codex",
+            test_commands=[],
+        )
+        workspace.status = status.value
+        workspace.updated_at = updated_at
+        workspace.compose_file_path = compose_file_path
+        if pr:
+            workspace.pr_url = "https://github.com/example/repo/pull/123"
+            workspace.pr_number = 123
+            workspace.pr_merge_sha = pr_merge_sha
+        await session.commit()
+        return workspace.id
+
+
+async def _set_workspace_gc_state(
+    session_factory: async_sessionmaker[AsyncSession],
+    workspace_id: str,
+    **values: object,
+) -> None:
+    async with session_factory() as session:
+        await session.execute(
+            update(Workspace)
+            .where(Workspace.id == workspace_id)
+            .values(**values)
+            .execution_options(synchronize_session=False)
+        )
+        await session.commit()
+
+
+async def _issue_gc_secret_lease(
+    session_factory: async_sessionmaker[AsyncSession],
+    workspace_id: str,
+    *,
+    now: datetime,
+) -> None:
+    async with session_factory() as session:
+        workspace = await WorkspaceRepository(session).get(workspace_id)
+        assert workspace is not None
+        await SecretLeaseRepository(session).issue_declared_leases(
+            workspace,
+            leases=[
+                SecretLeaseIssue(
+                    secret_name="api-token",
+                    kind="env",
+                    target="API_TOKEN",
+                    mode="ro",
+                    required=True,
+                    provider="env",
+                    ref_digest="sha256:" + "e" * 64,
+                    expires_at=now + timedelta(hours=1),
+                    issue_metadata={"profile": "gc", "declaration_index": 0},
+                )
+            ],
+            now=now,
+        )
+        await session.commit()
+
+
+def _write(path: Path, content: str) -> None:
+    path.parent.mkdir(parents=True, exist_ok=True)
+    path.write_text(content, encoding="utf-8")
+
+
+class _StaticRuntimeInspector:
+    def __init__(self, snapshot: RuntimeSnapshot) -> None:
+        self.snapshot = snapshot
+
+    async def inspect(self, compose_project_name: str | None) -> RuntimeSnapshot:
+        assert compose_project_name is not None
+        return self.snapshot
+
+
+@pytest.mark.unit
+async def test_completed_workspace_with_merge_sha_but_no_pr_metadata_is_preserved(
+    session_factory: async_sessionmaker[AsyncSession],
+    tmp_path: Path,
+) -> None:
+    now = datetime(2026, 4, 26, 12, tzinfo=UTC)
+    workspace_id = await _workspace(
+        session_factory,
+        status=WorkspaceStatus.completed,
+        updated_at=now - timedelta(hours=200),
+    )
+    async with session_factory() as session:
+        workspace = await session.get(Workspace, workspace_id)
+        assert workspace is not None
+        workspace.pr_merge_sha = "d" * 40
+        await session.commit()
+
+    plan = await plan_terminal_workspace_gc(
+        session_factory,
+        work_dir=tmp_path / "service",
+        min_age_hours=24,
+        now=now,
+    )
+
+    assert plan.candidates == []
+    assert len(plan.preserved) == 1
+    assert plan.preserved[0].workspace_id == workspace_id
+    assert plan.preserved[0].reason_code == "COMPLETED_WORKSPACE_WITHOUT_PR"
+
+
+@pytest.mark.unit
+async def test_completed_workspace_without_pr_metadata_does_not_consume_candidate_slot(
+    session_factory: async_sessionmaker[AsyncSession],
+    tmp_path: Path,
+) -> None:
+    now = datetime(2026, 4, 26, 12, tzinfo=UTC)
+    without_pr_id = await _workspace(
+        session_factory,
+        status=WorkspaceStatus.completed,
+        updated_at=now - timedelta(hours=300),
+        title="no pr metadata",
+    )
+    async with session_factory() as session:
+        workspace = await session.get(Workspace, without_pr_id)
+        assert workspace is not None
+        workspace.pr_merge_sha = "a" * 40
+        await session.commit()
+
+    with_pr_id = await _workspace(
+        session_factory,
+        status=WorkspaceStatus.completed,
+        updated_at=now - timedelta(hours=200),
+        title="has pr metadata",
+        pr=True,
+        pr_merge_sha="b" * 40,
+    )
+
+    plan = await plan_terminal_workspace_gc(
+        session_factory,
+        work_dir=tmp_path / "service",
+        min_age_hours=24,
+        limit=1,
+        now=now,
+    )
+
+    assert [c.workspace_id for c in plan.candidates] == [with_pr_id]
+    preserved_ids = {p.workspace_id for p in plan.preserved}
+    assert without_pr_id in preserved_ids
